Fix monitor
Some old versions of libraries were no longer being included, so
bumping the versions here.
Also traceview is no longer shipped as a separate executable. Monitor
was checking that the traceview executable existed on startup, since
there were a couple cases in which monitor invoked the standalone
traceview. However, in practice those places were never hit anymore
(which is reasonable since traceview is a part of monitor as well),
so switched to looking for "monitor" instead (since it's used to
determine the sdk directory).

Test: Just removing functionality. Also, this is for the
soon-to-be-removed eclipse-based tools, so hopefully no new tests are
needed.

Change-Id: Id52ea9e20532c740bd31d9cfbaef27aadd0713a7