Making "android" a wrapper for new tools
This will restore most of the avd-related functionality to "android".
The SDK functionality will be significantly affected, since the tools is
completely new, so a warning is printed (with acknowledgement required,
disableable via an argument or env var) that translation to the old
command syntax is best-effort.
Project functionality is not included.

Test: Added (manually run) bazel tests in sdk/files for the bash script,
and "test_android.bat" that you can run for the batch script.

Change-Id: Ie27e7c8ee2901609f46c91f3dac7e2c066fff98b