[SDK tools] Use a better type name for encoded characters
I define a name for a platform-dependent type to hold non-ASCII
character strings (wchar_t on Windows and char on Linux/Mac).

I previously called this WIDE_CHAR, but that is misleading
for the Linux case. My new choice is TCHAR, which is also
what Microsoft uses for this same purpose.

Bug: 113514423
Test: None: the change is purely cosmetic. (Famous last words!)
Change-Id: Icf99dac8e07ae9e25207a1683e588d227c0758c1